feat: componentize saved news header
