+++ src/components/SavedNews/SavedNews.css
@@ -5,52 +5,6 @@
   background-color: #f5f6f7;
 }
 
-.saved-news__header {
-  margin: 0;
-  padding: 0;
-  background-color: #fff;
-  height: auto;
-  padding: 40px 104px 56px;
-}
-
-.saved-news__section-title {
-  margin: 0;
-  padding: 0;
-  font-family: "Roboto";
-  font-size: 18px;
-  font-weight: 400;
-  line-height: 1.333;
-  color: #1a1b22;
-  opacity: 0.5;
-  margin-bottom: 28px;
-}
-
-.saved-news__legend_weight_bold {
-  font-weight: 700;
-}
-
-.saved-news__title {
-  margin: 0;
-  padding: 0;
-  font-family: "Roboto Slab";
-  font-size: 40px;
-  font-weight: 400;
-  line-height: 1.15;
-  color: #1a1b22;
-  white-space: pre-line;
-  margin-bottom: 30px;
-}
-
-.saved-news__legend {
-  margin: 0;
-  padding: 0;
-  font-family: "Roboto";
-  font-size: 18px;
-  font-weight: 400;
-  line-height: 1.333;
-  color: #1a1b22;
-}
-
 .saved-news__news {
   margin: 62px 104px;
   display: flex;
@@ -66,19 +20,6 @@
 @media screen and (max-width: 1080px) {
   .saved-news__news {
     margin: 32px 40px;
-  }
-  .saved-news__header {
-    padding: 24px 40px 32px;
-  }
-  .saved-news__section-title {
-    margin-bottom: 16px;
-    font-size: 14px;
-    line-height: 1.714;
-  }
-  .saved-news__title {
-    font-size: 30px;
-    line-height: 1.133;
-    margin-bottom: 16px;
   }
 }
 
@@ -103,18 +44,5 @@
 @media screen and (max-width: 410px) {
   .saved-news__news {
     margin: 24px 16px;
-  }
-  .saved-news__header {
-    padding: 8px 16px 32px;
-  }
-  .saved-news__section-title {
-    font-size: 12px;
-    line-height: 2;
-    margin-bottom: 16px;
-  }
-  .saved-news__title {
-    font-size: 30px;
-    line-height: 1.133;
-    margin-bottom: 50px;
   }
 }
