+++ src/components/SavedNewsHeader/SavedNewsHeader.css
@@ -1,0 +1,77 @@
+.saved-news-header {
+  margin: 0;
+  padding: 0;
+  background-color: #fff;
+  height: auto;
+  padding: 40px 104px 56px;
+}
+
+.saved-news-header__section-title {
+  margin: 0;
+  padding: 0;
+  font-family: "Roboto";
+  font-size: 18px;
+  font-weight: 400;
+  line-height: 1.333;
+  color: #1a1b22;
+  opacity: 0.5;
+  margin-bottom: 28px;
+}
+
+.saved-news-header__title {
+  margin: 0;
+  padding: 0;
+  font-family: "Roboto Slab";
+  font-size: 40px;
+  font-weight: 400;
+  line-height: 1.15;
+  color: #1a1b22;
+  white-space: pre-line;
+  margin-bottom: 30px;
+}
+
+.saved-news-header__legend_weight_bold {
+  font-weight: 700;
+}
+
+.saved-news-header__legend {
+  margin: 0;
+  padding: 0;
+  font-family: "Roboto";
+  font-size: 18px;
+  font-weight: 400;
+  line-height: 1.333;
+  color: #1a1b22;
+}
+
+@media screen and (max-width: 1080px) {
+  .saved-news-header__header {
+    padding: 24px 40px 32px;
+  }
+  .saved-news-header__section-title {
+    margin-bottom: 16px;
+    font-size: 14px;
+    line-height: 1.714;
+  }
+  .saved-news-header__title {
+    font-size: 30px;
+    line-height: 1.133;
+    margin-bottom: 16px;
+  }
+}
+
+@media screen and (max-width: 410px) {
+  .saved-news-header__header {
+    padding: 8px 16px 32px;
+  }
+  .saved-news-header__section-title {
+    font-size: 12px;
+    line-height: 2;
+    margin-bottom: 16px;
+  }
+  .saved-news-header__title {
+    font-size: 30px;
+    line-height: 1.133;
+    margin-bottom: 50px;
+  }
+}
